fix: move UserInfo to @types.ts
