+++ src/apis/auth/@types.ts
@@ -1,0 +1,12 @@
+export interface UserInfo {
+  id: string;
+  name: string;
+  permissions: {
+    admin: boolean;
+    blogAdmin: boolean;
+    blogWriter: boolean;
+    exhibition: boolean;
+    general: boolean;
+    reservation: boolean;
+  };
+}
