+++ src/apis/auth/user/index.ts
@@ -1,17 +1,5 @@
 import { AuthToken } from "../../@types";
-
-export interface UserInfo {
-  id: string;
-  name: string;
-  permissions: {
-    admin: boolean;
-    blogAdmin: boolean;
-    blogWriter: boolean;
-    exhibition: boolean;
-    general: boolean;
-    reservation: boolean;
-  };
-}
+import { UserInfo } from "../@types";
 
 export interface Methods {
   /**
